+++ ui/inputbox.c
@@ -28,7 +28,7 @@ COPYRIGHT 1993-1999 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 void strcins(char *s, int p, char c)
 {
 	int n;
-	for (n=strlen(s)-p; n>=0; n-- )
+	for (n = (int)strlen(s) - p; n >= 0; n--)
 		*(s+p+n+1) = *(s+p+n);   // Move everything over
 	*(s+p) = c;         // then insert the character
 }
